# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Smart Chess Coach API
  description: |
    Smart Chess Coach - AI-Powered Chess Learning Platform
    
    A comprehensive chess analysis and coaching platform that provides:
    - Game analysis with Stockfish engine
    - AI-powered personalized coaching
    - Player progression tracking
    - Mistake pattern analysis and learning
    - Voice-based coaching feedback
    - Journey dashboard with comprehensive stats
  version: 1.0.0
  contact:
    name: Smart Chess Coach Team
  license:
    name: Proprietary
servers:
  - url: https://api.smartchesscoach.com
    description: Production
  - url: http://localhost:8000
    description: Local development
...
paths:
...
  /api/games:
    get:
      tags:
        - Games
      summary: Get all games
      description: Retrieve all imported games for the authenticated user
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        '200':
          description: List of games
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Game'
        '401':
          description: Not authenticated
  /api/games/analyzed:
    get:
      tags:
        - Games
      summary: Get analyzed games with stats
      description: Retrieve all analyzed games with accuracy, blunders, and mistakes summary
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        '200':
          description: List of analyzed games with stats
          content:
            application/json:
              schema:
                type: object
                properties:
                  games:
                    type: array
                    items:
                      type: object
                      properties:
                        game_id:
                          type: string
                        opponent:
                          type: string
                        result:
                          type: string
                          enum: [win, loss, draw]
                        accuracy:
                          type: number
                          format: float
                        blunders:
                          type: integer
                        mistakes:
                          type: integer
                        best_moves:
                          type: integer
                        platform:
                          type: string
                  total:
                    type: integer
        '401':
          description: Not authenticated
  /api/games/blunders:
    get:
      tags:
        - Games
      summary: Get all blunders
      description: Retrieve all blunders from analyzed games with position and explanation
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        '200':
          description: List of blunders
          content:
            application/json:
              schema:
                type: object
                properties:
                  blunders:
                    type: array
                    items:
                      type: object
                      properties:
                        game_id:
                          type: string
                        move_number:
                          type: integer
                        move:
                          type: string
                        evaluation:
                          type: string
                          enum: [blunder, mistake]
                        fen:
                          type: string
                          description: FEN position before the blunder
                        feedback:
                          type: string
                        threat:
                          type: string
                  total:
                    type: integer
        '401':
          description: Not authenticated
  /api/games/best-moves:
    get:
      tags:
        - Games
      summary: Get all best moves
      description: Retrieve all best and excellent moves from analyzed games
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        '200':
          description: List of best moves
          content:
            application/json:
              schema:
                type: object
                properties:
                  best_moves:
                    type: array
                    items:
                      type: object
                      properties:
                        game_id:
                          type: string
                        move_number:
                          type: integer
                        move:
                          type: string
                        evaluation:
                          type: string
                          enum: [best, excellent, good]
                        fen:
                          type: string
                  total:
                    type: integer
        '401':
          description: Not authenticated
  /api/games/{game_id}:
    get:
      tags:
        - Games
      summary: Get specific game
      description: Retrieve detailed information for a specific game including PGN and termination reason
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - name: game_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Game details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '401':
          description: Not authenticated
        '404':
          description: Game not found
...
components:
  schemas:
...
    Game:
      type: object
      properties:
        game_id:
          type: string
        user_id:
          type: string
        platform:
          type: string
          enum: [chess.com, lichess]
        pgn:
          type: string
          description: PGN notation of the game
        white_player:
          type: string
        black_player:
          type: string
        result:
          type: string
          enum: ['1-0', '0-1', '1/2-1/2']
        time_control:
          type: string
          nullable: true
        date_played:
          type: string
          nullable: true
        opening:
          type: string
          nullable: true
        user_color:
          type: string
          enum: [white, black]
        imported_at:
          type: string
          format: date-time
        is_analyzed:
          type: boolean
      required:
        - game_id
        - user_id
        - platform
        - pgn
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_token
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Session token as Bearer token in Authorization header
security:
  - cookieAuth: []
  - bearerAuth: []